tbot/requirements.txt
requests
tinydb>=3.15.2
python-telegram-bot>=12.4.2,<13.0.0
